requirements.txt
Flask>=2.0.0
numpy>=1.19.0
tensorflow-cpu>=2.0.0
Pillow>=8.0.0
matplotlib>=3.3.0
Werkzeug>=2.0.0